CRX: Ensure that calcLength returns an uint32
Previously, if one of the inputs was undefined (because we read past the
end of the buffer), then the output could be NaN.

Or if d was somehow very large, then we could get a negative number.

Even though these do not occur in practice because such large CRX files
aren't supported, we fix the logic to avoid confusion.